revisao direta

+++ controllers/front/ListAllController.php
@@ -33,7 +33,7 @@ class ListAllControllerCore extends FrontController
 		parent::initContent();
 
 		$this->productSort();
-		$nbProducts = Product::getAllProducts($this->context->language->id, 0, 0, true, $this->orderBy, $this->orderWay),
+		$nbProducts = (int)Product::getAllProducts($this->context->language->id, 0, 0, true, $this->orderBy, $this->orderWay),
 		$this->pagination($nbProducts);
 
 		$this->context->smarty->assign(array(
